# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DroniAuth API",
        "description": "OAuth 2.0 Authorization Server. Provides authentication, user management, and OAuth client management endpoints.",
        "contact": {
            "email": "dev@droni.co"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://auth.droni.co",
            "description": "Production"
        },
        {
            "url": "http://localhost",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "List OAuth clients owned by the authenticated user",
                "operationId": "af1f3794367a5e843bf15e3a9033d911",
                "responses": {
                    "200": {
                        "description": "Client list with OAuth endpoint references",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clients": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OAuthClient"
                                            }
                                        },
                                        "oauthEndpoints": {
                                            "properties": {
                                                "authorize": {
                                                    "type": "string",
                                                    "example": "https://auth.droni.co/oauth/authorize"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "https://auth.droni.co/oauth/token"
                                                },
                                                "revoke": {
                                                    "type": "string",
                                                    "example": "https://auth.droni.co/oauth/tokens/{token_id}"
                                                },
                                                "userinfo": {
                                                    "type": "string",
                                                    "example": "https://auth.droni.co/api/user"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new OAuth client",
                "operationId": "15981a5da120d8b4ea1b91b05a35622b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "grant_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My App",
                                        "maxLength": 255
                                    },
                                    "grant_type": {
                                        "type": "string",
                                        "enum": [
                                            "authorization_code",
                                            "client_credentials"
                                        ]
                                    },
                                    "redirect_uris": {
                                        "description": "Newline-separated redirect URIs",
                                        "type": "string",
                                        "example": "https://myapp.com/callback"
                                    },
                                    "confidential": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true
                                    },
                                    "icon": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Client created – redirects to /clients with the plain secret in session flash"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/clients/{client}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Show an OAuth client and its authorized users",
                "operationId": "8fc07341eb8ece62ad4fdda3a30c7b25",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client details with list of authorized users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "client": {
                                            "$ref": "#/components/schemas/OAuthClient"
                                        },
                                        "authorizedUsers": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "format": "uri",
                                                        "nullable": true
                                                    },
                                                    "scopes": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "authorized_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "expires_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Client not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update an OAuth client",
                "operationId": "558832db987fc02c8a5107fb013891c9",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My App Updated",
                                        "maxLength": 255
                                    },
                                    "redirect_uris": {
                                        "description": "Newline-separated redirect URIs",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "logo": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true
                                    },
                                    "icon": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Client updated – redirects to /clients"
                    },
                    "404": {
                        "description": "Client not found or not owned by user"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete (revoke) an OAuth client",
                "operationId": "9102c26412072310db7ca015461a16cb",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Client deleted – redirects to /clients"
                    },
                    "404": {
                        "description": "Client not found or not owned by user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/clients/{client}/secret": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Regenerate the client secret",
                "operationId": "142470325154703af633ba13735bc2b5",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Secret regenerated – redirects to /clients with the new plain secret in session flash"
                    },
                    "404": {
                        "description": "Client not found or not owned by user"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "OAuthClient": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "My App"
                    },
                    "logo": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "icon": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "secret": {
                        "description": "Only present for confidential clients",
                        "type": "string",
                        "nullable": true
                    },
                    "redirect_uris": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        }
                    },
                    "grant_types": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "revoked": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "description": "Session cookie (web authentication)",
                "name": "droniauth_session",
                "in": "cookie"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth 2.0 Authorization Code flow issued by this server",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": {
                            "profile": "Read user profile (name, avatar)",
                            "email": "Read email address",
                            "roles": "Read roles and permissions",
                            "admin": "Full admin access"
                        }
                    }
                }
            }
        }
    },
...
}